Add files via upload
Bug fix: After 10 plots, (in a cylce of 10) the cyclical numbering, so restart with _L01, did not work.
Update: After going through a cycle, the NTFS attribute Date created would not update when a new log for that cycle_item(?) was made (So _L01 would now log a new / the 11th plot but the Date created would still be set at the date of the 1st plot). Now we check&delete if _L01 exists.
Update: In case you want to edit the batch file dynamically and need to add code prior to the call of chia.exe, you can. Just delete the number of charachters you added from the star-line (not sure about line-feeds, I assume 1 char).